A Simple Prototype
Not quite event channels but offers some of the functionality I imagined. In fact, some of the changes were so minor to make it work I felt some moderate joy. Is the ceremony to much? Currently it is too early to tell but this might not be the only prototype for this concept.
#12