Second Prototype
This prototype involves the MailboxProcessor type creating an asynchronous link between the events and what is done with them. On the surface it seems ideal. I wrote a transform function that creates a helpful tree with line numbers and exceptions to help create useful output such as the HTML tree prototype I designed. Performance wise we are looking at the struct events being passed into the mailbox and the many calls to parse production that are used which shouldn't be insanely heavy.
#12