Deepest Failure
The code was easy but the result left more questions. I need to review errors as a whole so the code itself can gain insights into the error rather than always relying on manual input from the programmer. The deepest failure seems logical at first but the result is not helpful.